# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/services/create": {
      "post": {
        "description": "Permite a un administrador registrar un nuevo servicio en el sistema con imágenes opcionales.",
        "operationId": "ServiceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "status",
                  "name",
                  "price",
                  "orderId"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "REPAIR",
                      "WARRANTY"
                    ],
                    "example": "REPAIR"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "IN_PROGRESS",
                      "COMPLETED",
                      "DELIVERED",
                      "PAID"
                    ],
                    "example": "IN_PROGRESS"
                  },
                  "name": {
                    "type": "string",
                    "example": "Reparación de motor"
                  },
                  "description": {
                    "type": "string",
                    "example": "Revisión y reparación completa del motor"
                  },
                  "price": {
                    "type": "number",
                    "example": 250
                  },
                  "orderId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Archivos de imagen (máx. 5, 5MB cada uno)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "El servicio ha sido creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos o faltantes"
          },
          "401": {
            "description": "No autorizado. Se requiere autenticación"
          },
          "403": {
            "description": "No tiene permisos para realizar esta acción"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Crear un nuevo servicio",
        "tags": [
          "Servicios"
        ]
      }
    },
    "/services/findAll": {
      "get": {
        "description": "Obtiene una lista paginada de servicios. Puede filtrarse por estado y tipo.",
        "operationId": "ServiceController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtrar servicios por estado",
            "schema": {
              "enum": [
                "IN_PROGRESS",
                "COMPLETED",
                "DELIVERED",
                "PAID"
              ],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filtrar servicios por tipo",
            "schema": {
              "enum": [
                "REPAIR",
                "WARRANTY"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de servicios obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Service"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado. Se requiere autenticación"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Obtener lista de servicios",
        "tags": [
          "Servicios"
        ]
      }
    },
    "/services/findOne/{id}": {
      "get": {
        "description": "Obtiene los detalles completos de un servicio específico mediante su ID.",
        "operationId": "ServiceController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID único del servicio (UUID)",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Servicio encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado. Se requiere autenticación"
          },
          "404": {
            "description": "No se encontró ningún servicio con el ID proporcionado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Obtener servicio por ID",
        "tags": [
          "Servicios"
        ]
      }
    },
    "/services/update/{id}": {
      "patch": {
        "description": "Actualiza los datos de un servicio existente. Solo actualiza los campos proporcionados.",
        "operationId": "ServiceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID único del servicio a actualizar",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Campos del servicio a actualizar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              },
              "examples": {
                "actualizarEstado": {
                  "summary": "Actualizar estado",
                  "value": {
                    "status": "COMPLETED"
                  }
                },
                "actualizarPrecio": {
                  "summary": "Actualizar precio",
                  "value": {
                    "price": 300
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Servicio actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado. Se requiere autenticación"
          },
          "403": {
            "description": "No tiene permisos para realizar esta acción"
          },
          "404": {
            "description": "No se encontró el servicio especificado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Actualizar un servicio",
        "tags": [
          "Servicios"
        ]
      }
    },
    "/services/remove/{id}": {
      "delete": {
        "description": "Elimina permanentemente un servicio del sistema. Esta acción no se puede deshacer.",
        "operationId": "ServiceController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID único del servicio a eliminar",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Servicio eliminado exitosamente"
          },
          "401": {
            "description": "No autorizado. Se requiere autenticación"
          },
          "403": {
            "description": "No tiene permisos para realizar esta acción"
          },
          "404": {
            "description": "No se encontró el servicio especificado"
          },
          "409": {
            "description": "No se puede eliminar el servicio porque tiene registros asociados"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Eliminar un servicio",
        "tags": [
          "Servicios"
        ]
      }
    }
  },
  "info": {
    "title": "API Documentation",
    "description": "Documentación de la API de la aplicación",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Ingrese el token JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "Service": {
        "type": "object",
        "properties": {}
      },
      "UpdateServiceDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "REPAIR",
              "WARRANTY"
            ],
            "default": "REPAIR",
            "description": "Tipo de servicio",
            "example": "REPAIR"
          },
          "status": {
            "type": "string",
            "enum": [
              "IN_PROGRESS",
              "COMPLETED",
              "DELIVERED",
              "PAID"
            ],
            "default": "IN_PROGRESS",
            "description": "Estado del servicio",
            "example": "IN_PROGRESS"
          },
          "name": {
            "type": "string",
            "description": "Nombre del servicio",
            "example": "Reparación de motor"
          },
          "description": {
            "type": "string",
            "description": "Descripción detallada del servicio",
            "example": "Revisión y reparación completa del motor"
          },
          "photoUrls": {
            "default": [],
            "description": "URLs de las fotos del servicio",
            "example": [
              "https://ejemplo.com/foto1.jpg",
              "https://ejemplo.com/foto2.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "price": {
            "type": "number",
            "description": "Precio del servicio",
            "example": 250.5
          },
          "orderId": {
            "type": "string",
            "description": "ID de la orden asociada",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        }
      }
    }
  }
}